Remote Server API Authentication for sync_files
