+++ server/server/src/static_files.rs
@@ -5,6 +5,7 @@ use std::sync::Mutex;
 
 use actix_files as fs;
 use actix_multipart::Multipart;
+
 use actix_web::error::InternalError;
 use actix_web::http::header::{ContentDisposition, DispositionParam, DispositionType};
 use actix_web::http::StatusCode;
@@ -18,6 +19,7 @@ use serde::{Deserialize, Serialize};
 
 use repository::sync_file_reference_row::SyncFileReferenceRow;
 
+use service::auth_data::AuthData;
 use service::plugin::plugin_files::{PluginFileService, PluginInfo};
 use service::plugin::validation::ValidatedPluginBucket;
 use service::service_provider::ServiceProvider;
@@ -27,6 +29,7 @@ use service::sync::file_sync_driver::get_sync_settings;
 use service::sync::file_synchroniser::FileSynchroniser;
 use util::is_central_server;
 
+use crate::authentication::validate_cookie_auth;
 use crate::middleware::limit_content_length;
 
 // this function could be located in different module
@@ -164,7 +167,16 @@ async fn delete_sync_file(
     settings: Data<Settings>,
     service_provider: Data<ServiceProvider>,
     path: web::Path<(String, String, String)>,
+    request: HttpRequest,
+    auth_data: Data<AuthData>,
 ) -> Result<HttpResponse, Error> {
+    validate_cookie_auth(request.clone(), &auth_data).map_err(|_err| {
+        InternalError::new(
+            "You must be logged in to delete files",
+            StatusCode::INTERNAL_SERVER_ERROR,
+        )
+    })?;
+
     let (table_name, record_id, file_id) = path.into_inner();
     let static_file_category = StaticFileCategory::SyncFile(table_name, record_id);
 
@@ -202,8 +214,18 @@ async fn upload_sync_file(
     settings: Data<Settings>,
     service_provider: Data<ServiceProvider>,
     path: web::Path<(String, String)>,
+    request: HttpRequest,
+    auth_data: Data<AuthData>,
 ) -> Result<HttpResponse, Error> {
-    // TODO Authorization
+    // For now, we just check that the user is authenticated
+    // In future we might want to check that the user has access to the record
+    // Access to the file UUID should normally only be exposed to users with access from the frontend
+    validate_cookie_auth(request.clone(), &auth_data).map_err(|_err| {
+        InternalError::new(
+            "You need to be logged in",
+            StatusCode::INTERNAL_SERVER_ERROR,
+        )
+    })?;
 
     let db_connection = service_provider
         .connection()
@@ -262,9 +284,20 @@ async fn sync_files(
     settings: Data<Settings>,
     service_provider: Data<ServiceProvider>,
     path: web::Path<(String, String, String)>,
+    auth_data: Data<AuthData>,
 ) -> Result<HttpResponse, Error> {
+    // For now, we just check that the user is authenticated
+    // In future we might want to check that the user has access to the record
+    // Access to the file UUID should normally only be exposed to users with access from the frontend
+    validate_cookie_auth(req.clone(), &auth_data).map_err(|_err| {
+        InternalError::new(
+            "You need to be logged in",
+            StatusCode::INTERNAL_SERVER_ERROR,
+        )
+    })?;
+
     let service = StaticFileService::new(&settings.server.base_dir)
-        .map_err(|err| InternalError::new(err, StatusCode::INTERNAL_SERVER_ERROR))?;
+        .map_err(|err| InternalError::new(err, StatusCode::FORBIDDEN))?;
 
     let (table_name, parent_record_id, file_id) = path.into_inner();
